% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute.sensitivity.function.coefs.R
\name{compute.sensitivity.function.coefs}
\alias{compute.sensitivity.function.coefs}
\title{Compute the coefs of the (multilinear) N-way
sensitivity function}
\usage{
compute.sensitivity.function.coefs(grain.domain, nway.info, unique.params.list,
  param.functions.list, e.mat, h.mat)
}
\arguments{
\item{grain.domain}{}

\item{nway.info}{}

\item{unique.params.list}{}

\item{param.functions.list}{}

\item{e.mat}{}

\item{h.mat}{}
}
\value{
a vector
}
\description{
Compute the coefs of the (multilinear) N-way
sensitivity function
}
\details{
none
}
\examples{
Coming soon.

\dontrun{

}
}

