% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.lower.triangle.R
\name{extract.lower.triangle}
\alias{extract.lower.triangle}
\title{Extract the lower triangle of elements in a square matrix}
\usage{
extract.lower.triangle(sqmat)
}
\arguments{
\item{sqmat}{}
}
\value{
an array
}
\description{
Extract the lower triangle of elements in a square matrix
}
\details{
Handy to help check if any of the rows in the perturbed prob mat
are "to close" to each other
}
\examples{
Coming soon.

\dontrun{

}
}

