% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fool.compileCPT.R
\name{fool.compileCPT}
\alias{fool.compileCPT}
\title{Utility to print out a gRain CPT from it's network}
\usage{
fool.compileCPT(vpar.list, node.names)
}
\arguments{
\item{vpar.list}{a list of the child node and its parents.}

\item{node.names}{string versions of the names for the nodes listed in vpar.list.}
}
\value{
a \code{parray}
}
\description{
A handy function to help print out gRain CPTs
}
\details{
This utility will compile a CPT into the necessary
parray (gRbase) form needed for gRain2HuginCPT to print it out. The trick is that the
parent nodes and node names need to be included.
EG: fool.compileCPT(list(bg.seeds, garnet, loc), c("bg.seeds","garnet", "loc"))[[1]]
spits out the parray for bg.seeds. bg.seeds has parent nodes garnet and loc.
}
\examples{
Coming soon.

\dontrun{

}
}

