% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.gRainCPT.R
\name{get.gRainCPT}
\alias{get.gRainCPT}
\title{Get gRain CPT out of a gRain domain (network)}
\usage{
get.gRainCPT(gRain.domain, node.name)
}
\arguments{
\item{gRain.domain}{a gRain network}

\item{node.name}{the name of the node as a character string}
}
\value{
a gRain \code{cptable}
}
\description{
Handy wrapper to grab a gRain CPT by name out of a gRain network.
}
\details{
Handy wrapper to grab a gRain CPT by name (as a string) out of a gRain network.
}
\examples{
Coming soon.

\dontrun{

}
}

