% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gRain.pr.index.R
\name{gRain.pr.index}
\alias{gRain.pr.index}
\title{Get the index of a specified state probability
in a gRain CPT.}
\usage{
gRain.pr.index(gRain.CPT, prior.nodeQ = NULL, nslist = NULL)
}
\arguments{
\item{gRain.CPT}{}

\item{prior.nodeQ=NULL}{}

\item{nslist=NULL}{}
}
\value{
an array of row indices
}
\description{
Get the index of a specified state probability
in a gRain CPT.
}
\details{
nslist is the variable specification list. They can
be specified in any order.
NOTE: A gRain.CPT is just an array or parray
}
\examples{
Coming soon.

\dontrun{

}
}

