% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gRain.prior.nodeQ.R
\name{gRain.prior.nodeQ}
\alias{gRain.prior.nodeQ}
\title{Test a gRain CPT to see if is a prior node}
\usage{
gRain.prior.nodeQ(gRain.CPT)
}
\arguments{
\item{gRain.CPT}{A gRain formatted CPT}
}
\value{
TRUE or FALSE
}
\description{
Test a gRain CPT to see if is a prior node
}
\details{
The function is an internal routine, but may be handy
for some applications.
}
\examples{
Coming soon.

\dontrun{

}
}

