% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gRain2HuginCPT.R
\name{gRain2HuginCPT}
\alias{gRain2HuginCPT}
\title{Convert a gRain format CPT to a Hugin format CPT}
\usage{
gRain2HuginCPT(gRain.CPT, prior.nodeQ = NULL)
}
\arguments{
\item{gRain.CPT}{A gRain formatted CPT}

\item{prior.nodeQ}{switch to indicate if CPT is a prior node.}
}
\value{
A Hugin format CPT.
}
\description{
Convert a gRain format CPT to a Hugin format CPT
}
\details{
#Convert a gRain format CPT to a Hugin format CPT.
gRain format is a (multi) table. Hugin format
is a case-frequency data.frame.
NOTE: A gRain.CPT is just an array or parray. See gRbase documentation.
}
\examples{
Coming soon.

\dontrun{

}
}

