% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grain2PrettyCPT.R
\name{gRain2PrettyCPT}
\alias{gRain2PrettyCPT}
\title{Convert a gRain format CPT to a pretty Hugin format CPT}
\usage{
gRain2PrettyCPT(gRain.CPT, prior.nodeQ = NULL)
}
\arguments{
\item{gRain.CPT}{A gRain formatted CPT}

\item{prior.nodeQ}{switch to indicate if CPT is a prior node.}
}
\value{
A row-wise CPT with nicely formatted node-state strings.
}
\description{
Convert a gRain format CPT to a Hugin format CPT, but with nicely
formatted strings indicating the conditional probability state (row of the Hugin CPT).
}
\details{
#Convert a gRain format CPT to a Hugin format CPT.
gRain format is a (multi) table. Hugin format
is a case-frequency data.frame. This function nicely formats the rows.
}
\examples{
Coming soon.

\dontrun{

}
}

