\name{gRmu-package}
\alias{gRmu-package}
\alias{gRmu}
\docType{package}
\title{
Utilities for Graphical Models in gRain
}
\description{
Functions and GUI tools to interact with/view graphical models in gRain
}
\details{
\tabular{ll}{
Package: \tab gRmu\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-08\cr
License: \tab GPL-3\cr
}
Handy functions to interact with graphical models made with or pulled into gRain.
}
\author{
Nick Petraco and Tanya Napolitano

Maintainer: <npetraco@gmail.com>
}
\references{
http://cran.r-project.org/web/packages/gRain/index.html\cr
http://reasoning.cs.ucla.edu/samiam/\cr
https://dslpitt.org/genie/\cr
http://www.hugin.com/\cr
}
