% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initialize.gRainCPT.R
\name{initialize.gRainCPT}
\alias{initialize.gRainCPT}
\title{Utility to initialize a gRain CPT.}
\usage{
initialize.gRainCPT(vpar.list, child.node.levels = NULL, printQ = FALSE)
}
\arguments{
\item{vpar.list}{a list of the child node and its parents.}

\item{child.node.levels}{input levels for the node.}

\item{printQ}{switch to print the CPT created.}
}
\value{
A \code{cptable} object (a list).
}
\description{
A function to create and initialize a gRain CPT. Handy when
building a large network by hand in R.
}
\details{
This utility will format a vpar.list and input levels for the node (child.node.levels)
into a gRain CPT with the cptable function.  For the definition of vpar.list cf. help page
of cptable (gRain).
}
\examples{
Coming soon.

\dontrun{

}
}

