% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.A.R
\name{make.A}
\alias{make.A}
\title{Generate an A matrix from the perturbed prob mat}
\usage{
make.A(pert.prob.mat)
}
\arguments{
\item{pert.prob.mat}{}
}
\value{
a matrix
}
\description{
Generate an A matrix from the perturbed prob mat
}
\details{
Need A for y=Ax
}
\examples{
Coming soon.

\dontrun{

}
}

