% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.perterbed.prob.mat.R
\name{make.perterbed.prob.mat}
\alias{make.perterbed.prob.mat}
\title{Generate a martix of perturbed probs for doing
n-way senstivity analysis}
\usage{
make.perterbed.prob.mat(num.states)
}
\arguments{
\item{num.states}{}
}
\value{
a vector
}
\description{
Generate a martix of perturbed probs for doing
n-way senstivity analysis
}
\details{
NOTE: here n == num.states
}
\examples{
Coming soon.

\dontrun{

}
}

