% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.perterbed.prob.mat2.R
\name{make.perterbed.prob.mat2}
\alias{make.perterbed.prob.mat2}
\title{Make perterbed probability matrix for sensitivity analysis}
\usage{
make.perterbed.prob.mat2(num.states)
}
\arguments{
\item{num.states}{}
}
\value{
a matrix
}
\description{
Make perterbed probability matrix for sensitivity analysis
}
\details{
none
}
\examples{
Coming soon.

\dontrun{

}
}

