% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/merge.probs.R
\name{merge.probs}
\alias{merge.probs}
\title{Merge (interlace) probabilities (beliefs) with their complements}
\usage{
\method{merge}{probs}(prob.vec)
}
\arguments{
\item{prob.vec}{vector of probabilites to complement and merge}
}
\description{
Merge (interlace) probabilities with their complements to make a
(gRain or Hugin) ordered prob vector for a CPT.
}
\details{
The function is handy for sensitivity analysis.
NOTE: ONLY FOR BINARY NODES!
}
\examples{
Coming soon.

\dontrun{

}
}

