% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mod.and.set.hugin.cpt.R
\name{mod.and.set.hugin.cpt}
\alias{mod.and.set.hugin.cpt}
\title{Set beliefs (probabilities) of states for a Hugin CPT}
\usage{
mod.and.set.hugin.cpt(hugin.domain, node.name, prob.vec, printQ = FALSE)
}
\arguments{
\item{hugin.domain}{RHugin loaded Hugin domain}

\item{node.name}{name of the node (CPT)}

\item{prob.vec}{probability vector to set}

\item{printQ}{switch to print the CPT}
}
\value{
Nothing.
}
\description{
\code{mod.and.set.hugin.cpt} Set Prs of states for a Hugin CPT.
}
\details{
Modify and set a CPT in a Hugin domain.
}
\examples{
Coming soon.

\dontrun{

}
}

