% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monomial.code.R
\name{monomial.code}
\alias{monomial.code}
\title{Coupe/Jensen code for a monomial in the multilinear function}
\usage{
monomial.code(i)
}
\arguments{
\item{i}{}
}
\value{
a string
}
\description{
Coupe/Jensen code for a monomial in the multilinear function
}
\details{
none
}
\examples{
Coming soon.

\dontrun{

}
}

