% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monomial.combns.R
\name{monomial.combns}
\alias{monomial.combns}
\title{Canonically orderd indices of monomials excluding the constant term}
\usage{
monomial.combns(nn)
}
\arguments{
\item{nn}{}
}
\value{
a vector
}
\description{
Canonically orderd indices of monomials excluding the constant term
}
\details{
none
}
\examples{
Coming soon.

\dontrun{

}
}

