% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/one.way.sensitivity.metrics.R
\name{one.way.sensitivity.metrics}
\alias{one.way.sensitivity.metrics}
\title{Compute local sensitivity and vertex proximities
of parameters in a one-way sensitivity analysis}
\usage{
one.way.sensitivity.metrics(x0, one.waycoef.mat)
}
\arguments{
\item{x0}{}

\item{one.waycoef.mat}{}
}
\value{
an array
}
\description{
Compute local sensitivity and vertex proximities
of parameters in a one-way sensitivity analysis
}
\details{
Cf. van der Gaag, Renooij and Coupe
}
\examples{
Coming soon.

\dontrun{

}
}

