% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/propagate.perterbed.probs.R
\name{propagate.perterbed.probs}
\alias{propagate.perterbed.probs}
\title{Compute necessary probabilities to solve for
sensitivity function coefs.}
\usage{
propagate.perterbed.probs(grain.domain, pert.prob.mat, nway.info,
  unique.params.list, param.functions.list, e.mat, h.mat)
}
\arguments{
\item{grain.domain}{}

\item{pert.prob.mat}{}

\item{nway.info}{}

\item{unique.params.list}{}

\item{param.functions.list}{}

\item{e.mat}{}

\item{h.mat}{}
}
\value{
an array
}
\description{
Compute necessary probabilities to solve for
sensitivity function coefs.
}
\details{
Cf. Coupe, Jensen et al N-way Sensititvity analysis
}
\examples{
Coming soon.

\dontrun{

}
}

