% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.cpt.R
\name{read.cpt}
\alias{read.cpt}
\title{Read in a CPT to and convert to Hugin order if not already done so}
\usage{
read.cpt(fpath)
}
\arguments{
\item{fpath}{}
}
\value{
an array
}
\description{
Read in a CPT to and convert to Hugin order if not
already done so
}
\details{
NOTE, If coplements were not originally in the file, they WILL NOT
BE replaced. One has to do that manually
}
\examples{
Coming soon.

\dontrun{

}
}

