% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove.sam.tabs.R
\name{remove.sam.tabs}
\alias{remove.sam.tabs}
\title{Remove extra spaces from Hugin .net file generated by SamIam.}
\usage{
remove.sam.tabs(fpath)
}
\arguments{
\item{fpath}{The path to the .net file}
}
\value{
Nothing.
}
\description{
\code{remove.sam.tabs} Removes extra spaces from SamIam .net file.
}
\details{
Remove the tabs put into a .net file saved from SamIam. The file is
re-written. The tabs make loadHuginNet in gRain choke.
}
\examples{
Coming soon.

\dontrun{
 remove.sam.tabs("path/to/file/blah.net")
}
}

