% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set.gRainCPT.R
\name{set.gRainCPT}
\alias{set.gRainCPT}
\title{Higher level utility function to set probabilities (beliefs) in an
initialized or existing gRain CPT}
\usage{
set.gRainCPT(node.cpt, prob.vec, complement.prob.vecQ = FALSE)
}
\arguments{
\item{node.cpt}{a gRain CPT}

\item{prob.vec}{a vector of beliefs}

\item{complement.prob.vecQ}{switch to interlace the complement of \code{prob.vec} if required.}
}
\value{
A \code{cptable} object (a list).
}
\description{
Set beliefs in a gRain CPT. Handy for sensitivity analysis.
}
\details{
Set beliefs (probabilities) in a gRain CPT. Handy for sensitivity analysis.
Function is really just a wrapper for boilerplate code to input beliefs into an
existing \code{cptable}. NOTE: The \code{complement.prob.vecQ} flag should only be switch on if the
node is BINARY.
}
\examples{
Coming soon.

\dontrun{

}
}

