% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.cpt.R
\name{write.cpt}
\alias{write.cpt}
\title{Write a CPT to file in Hugin order}
\usage{
write.cpt(grain.cpt, prior.node, fpath, remove.complimentQ = F,
  freq.col = "last", row.numsQ = T)
}
\arguments{
\item{grain.cpt}{}

\item{prior.node}{}

\item{fpath}{}

\item{remove.complimentQ}{}

\item{freq.col}{}

\item{row.numsQ}{}
}
\value{
nothing
}
\description{
Write a CPT to file in Hugin order
}
\details{
none
}
\examples{
Coming soon.

\dontrun{

}
}

